requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
numpy==1.24.4
    # via pandas
pandas==2.0.3
    # via exabyte-standata (pyproject.toml)
python-dateutil==2.8.2
    # via pandas
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via exabyte-standata (pyproject.toml)
six==1.16.0
    # via python-dateutil
tzdata==2023.3
    # via pandas
